Promote PetRepositoryImpl to commonMain
